Restructuring
Restructuring package so Image class contains fewer functions that act on itself.
